+++ src/stories/navigation/Navbar.test.tsx
@@ -1,16 +1,38 @@
 import { fireEvent, screen, render } from "@testing-library/react";
-import { Navbar } from "stories/navigation/Navbar";
+import { Navbar, NavLinks, NavLink } from "stories/navigation/Navbar";
 
 describe("Navbar", () => {
+  it("should not display logo by default", () => {
+    global.innerWidth = 600;
+    render(<Navbar />);
+    expect(screen.getByAltText("Home")).toBeInTheDocument();
+  });
+
+  it("should not display logo if null", () => {
+    global.innerWidth = 600;
+    render(<Navbar logo={null} />);
+    expect(screen.queryByAltText("Home")).not.toBeInTheDocument();
+  });
+});
+
+describe("Navbar Links", () => {
   it("should display hamburger menu on narrow displays", () => {
     global.innerWidth = 600;
-    render(<Navbar links={[{ route: "/", label: "Proposals" }]} />);
+    render(
+      <NavLinks>
+        <NavLink>Proposals</NavLink>
+      </NavLinks>,
+    );
     expect(screen.getByRole("button", { name: /open menu/i })).toBeInTheDocument();
   });
 
   it("should display menu items when hamburger menu is clicked", () => {
     global.innerWidth = 600;
-    render(<Navbar links={[{ route: "/", label: "Proposals" }]} />);
+    render(
+      <NavLinks>
+        <NavLink>Proposals</NavLink>
+      </NavLinks>,
+    );
     const menu = screen.getByRole("button", { name: /open menu/i });
     fireEvent.click(menu);
 
@@ -19,19 +41,11 @@ describe("Navbar", () => {
 
   it("should render links properly", () => {
     global.innerWidth = 600;
-    render(<Navbar links={[{ route: "/", label: "Proposals" }]} />);
+    render(
+      <NavLinks>
+        <NavLink>Proposals</NavLink>
+      </NavLinks>,
+    );
     expect(screen.getByText("Proposals")).toBeInTheDocument();
-  });
-
-  it("should not display logo by default", () => {
-    global.innerWidth = 600;
-    render(<Navbar links={[{ route: "/", label: "Proposals" }]} />);
-    expect(screen.getByAltText("Home")).toBeInTheDocument();
-  });
-
-  it("should not display logo if null", () => {
-    global.innerWidth = 600;
-    render(<Navbar logo={null} links={[{ route: "/", label: "Proposals" }]} />);
-    expect(screen.queryByAltText("Home")).not.toBeInTheDocument();
   });
 });